[NUI] Add AnimatedVectorImageView
